agent-bringo-voice-stable/requirements.txt
# Core dependencies
google-cloud-aiplatform>=1.40.0
google-cloud-bigquery>=3.14.0
google-cloud-bigquery-storage>=2.24.0
db-dtypes>=1.1.1
pandas>=2.1.0
pyarrow>=14.0.0
google-cloud-storage>=2.14.0
google-cloud-secret-manager>=2.16.0  # Secure API key storage

# Vertex AI
vertexai>=1.40.0
google-cloud-discoveryengine>=0.11.0  # For Ranking API
google-genai>=0.1.0

# Image processing
Pillow>=10.0.0
requests>=2.31.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
tenacity>=8.2.3
tqdm>=4.66.1
python-dotenv>=1.0.0
streamlit>=1.42.0
Authlib>=1.3.2

# Optional: Development
pytest>=7.4.0
black>=23.12.0
ruff>=0.1.0

# Scraping & Automation
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0

# Database
psycopg2-binary>=2.9.9  # PostgreSQL adapter for Python
